[COS-6744] sli to products
